requirements.txt
# Requirements
pillow